user_service/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.3.6
certifi==2025.4.26
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
confluent-kafka==2.3.0
distlib==0.3.9
Django==5.2.1
django-environ==0.12.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.18.0
gunicorn==23.0.0
kombu==5.5.4
packaging==25.0
pbr==6.1.1
pipenv==2025.0.3
pipenv-to-requirements==0.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.0.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
